Moved Plugin class from cdap-api to cdap-app-fabric
