+++ cdap-app-fabric/src/main/java/co/cask/cdap/app/ApplicationSpecification.java
@@ -31,7 +31,7 @@ import co.cask.cdap.api.spark.SparkSpecification;
 import co.cask.cdap.api.worker.WorkerSpecification;
 import co.cask.cdap.api.workflow.Workflow;
 import co.cask.cdap.api.workflow.WorkflowSpecification;
-import co.cask.cdap.internal.artifact.Plugin;
+import co.cask.cdap.internal.app.Plugin;
 import co.cask.cdap.internal.dataset.DatasetCreationSpec;
 import co.cask.cdap.proto.Id;
